Perf: gather-free EM read, bf16 weights, remove sync points (3.9x speedup)
- Gather-free EM read: replace topk+gather+expand with masked dense
  attention. Eliminates GatherBackward (20ms) and ExpandBackward (15ms)
  from backward pass by using top-k masking instead of explicit gather.
- bf16 model weights on CUDA: store params in bf16 instead of fp32,
  eliminating 246 per-operation autocast casts (6.7ms CUDA saved).
- Remove CPU-GPU sync points: .any() on CUDA tensors forces
  cudaStreamSynchronize (14.5ms). Use unconditional masking instead.
- Remove redundant .to() dtype conversions in PM/EM read/write/reset.
- Remove is_initialized() branch checks from hot path (column, model).

Result: 7,263 -> 28,200 tok/s compiled (3.9x), VRAM 17.4 -> 4.5 GB.

+++ src/model/utils.py
@@ -86,7 +86,7 @@ class StateMixin:
                 expanded = mask
                 for _ in range(t.dim() - 1):
                     expanded = expanded.unsqueeze(-1)
-                setattr(self, name, t * (~expanded).to(t.dtype))
+                setattr(self, name, t * ~expanded)
 
     def state_dict_runtime(self) -> Dict[str, Optional[Tensor]]:
         """Serialize runtime state for checkpointing."""
